\name{trcpp-package}
\alias{trcpp-package}
\alias{trcpp}
\docType{package}
\title{
\packageTitle{trcpp}
}
\description{
\packageDescription{trcpp}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{trcpp}
\packageIndices{trcpp}
~~ An overview of how to use the package, including ~~
~~ the most important functions ~~
}
\author{
\packageAuthor{trcpp}

Maintainer: \packageMaintainer{trcpp}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, ~~
~~ from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
